src/__init__.py
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    print("Warning: python-dotenv not installed. Environment variables from .env file won't be loaded automatically.")
    print("Install it with: pip install python-dotenv")
