pyproject.toml
[project]
name = "rag-project"
version = "0.2.0"
description = "Corrective RAG System with self-correction and web search fallback"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "duckduckgo-search>=7.1.2",
    "fastapi>=0.118.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.37.0",
]
